fix(face): unify encoding dtype to float32 for W3 consistency
- face_helper.face_encodings() 返回 np.float32 而非 np.float64，
  与 dlib 内部 compute_face_descriptor 的输出 + FaceEncoding 列注释一致
- 加单测 test_face_encodings_dtype_is_float32 锁定，防止回归

影响范围：W3 face_service 序列化/比对链路
  之前 float64 → 序列化为 float32 → 还原后量纲变化，欧氏距离偏大
  现在 float32 → 序列化 float32 → 还原后一致

+++ src/utils/face_helper.py
@@ -102,7 +102,9 @@ def face_encodings(image_bgr: np.ndarray,
         # 转 dlib.rectangle
         rect = dlib.rectangle(left, top, right, bottom)
         shape = _sp(rgb, rect)               # 68 关键点
-        encoding = np.array(_facerec.compute_face_descriptor(rgb, shape), dtype=np.float64)
+        # 统一用 float32：dlib 内部 compute_face_descriptor 返回的就是 float32，
+        # face.py:16 列注释也写 "128维float32"，避免 W3 序列化/比对时量纲不一致
+        encoding = np.array(_facerec.compute_face_descriptor(rgb, shape), dtype=np.float32)
         encodings.append(encoding)
     return encodings
